Добавлена связь One To Many между Объявлением и Категорией

+++ src/main/java/ru/solarlab/study/dto/CategoryDto.java
@@ -23,20 +23,32 @@ public class CategoryDto {
 
     @PositiveOrZero
     @Schema(description = "Идентификатор")
-    public Integer id;
+    public long id;
 
+    /**
+     * Дата и время создания категории
+     */
     @NotBlank
     @Schema(description = "Дата и время создания категории")
     public OffsetDateTime createdAt;
 
+    /**
+     * Дата и время обновления категории
+     */
     @Schema(description = "Дата и время обновления категории")
     public OffsetDateTime updatedAt;
 
+    /**
+     * Имя категории
+     */
     @NotBlank
     @CapitalLetter
     @Schema(description = "Имя категории")
     public String name;
 
+    /**
+     * Статус категории
+     */
     @NotNull
     @Schema(description = "Статус")
     public CategoryStatus status;
